requirements.txt
Flask==3.0.0
markdown==3.5.1
python-frontmatter==1.1.0
gunicorn==22.0.0
